[comp 512][a1][client][refactor][5/7] move client and commands to commons, rmi to rmi
## Summary:
we want to refactor the starter code in a way such that it is easy to drop in a new tcp server/client in the future

we want to achieve something similar to the server side, where we have one Client that can either accept a TCP or RMI client and have an adapter (or proxy here) that deals with the networking differences

this diff:
move the common components (the client and commands) to their own directory

move the rmi components to the rmi directory

## Test Plan:
it builds at top of stack
